Tests/LibWeb/Text/expected/anchor-parse-nested-invalid.txt

no space before garbage: 
valid nested: anchor(--foo top, anchor(--bar bottom))
trailing whitespace: anchor(--foo top, anchor(--bar bottom))
